[Behat] Compound Payment and PaymentMethod transformer
