+++ src/Sylius/Behat/Resources/config/suites/ui/checkout/checkout.yml
@@ -10,7 +10,7 @@ default:
                 - sylius.behat.context.transform.addressing
                 - sylius.behat.context.transform.country
                 - sylius.behat.context.transform.lexical
-                - sylius.behat.context.transform.payment_method
+                - sylius.behat.context.transform.payment
                 - sylius.behat.context.transform.product
                 - sylius.behat.context.transform.zone
                 - sylius.behat.context.transform.shared_storage
