+++ src/Sylius/Behat/Resources/config/suites/ui/checkout/paying_for_order.yml
@@ -9,7 +9,7 @@ default:
 
                 - sylius.behat.context.transform.addressing
                 - sylius.behat.context.transform.lexical
-                - sylius.behat.context.transform.payment_method
+                - sylius.behat.context.transform.payment
                 - sylius.behat.context.transform.product
                 - sylius.behat.context.transform.shared_storage
