+++ src/Sylius/Behat/Resources/config/suites/ui/payment/managing_payment_methods.yml
@@ -13,7 +13,7 @@ default:
                 - sylius.behat.context.setup.security
 
                 - sylius.behat.context.transform.locale
-                - sylius.behat.context.transform.payment_method
+                - sylius.behat.context.transform.payment
                 - sylius.behat.context.transform.shared_storage
 
                 - sylius.behat.context.ui.admin.managing_payment_methods
